+++ src/disable-toolbar.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 // Remove links from admin toolbar.
 add_action('admin_bar_menu', function ($menu) {
-    $items = get_theme_support('plate-toolbar');
+    $items = get_theme_support('plate-disable-toolbar');
 
     foreach (reset($items) as $item) {
         $menu->remove_node($item);